Reverted beams test to test_design
